+++ src/index.ts
@@ -2,4 +2,5 @@ import crossp from './crossp';
 
 // Public package API
 export { crossp } from './crossp';
+export { limits } from './limits';
 export default crossp;