+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [Version 0.2.0](https://github.com/csquare-ai/crossp/releases/tag/0.2.0)
+
+- **feat**: delegate `RangeToken` calculations to the `bignumber.js` library
+- **fix**: rounding issues ([#1](https://github.com/csquare-ai/crossp/issues/1), [#3](https://github.com/csquare-ai/crossp/issues/3))
+- **tests**: refine tests (remove `it.each` when possible), add [#3](https://github.com/csquare-ai/crossp/issues/3) test case
+
 ## [Version 0.1.2 - Fix types directory](https://github.com/csquare-ai/crossp/releases/tag/0.1.2)
 
 - **fix**: set types directory to `dist/types`